throttler: Use integer value when deciding to do a minimal increment.
I saw instances where it should have incremented the rate but it didn't.

e.g.:
oldRate = 24
newRate = 24

My guess is that "newRate" was actually something like 24.01 and therefore "if newRate <= float64(oldRate)" didn't get triggered.

+++ go/vt/throttler/max_replication_lag_module.go
@@ -359,7 +359,7 @@ func (m *MaxReplicationLagModule) increaseRate(r *result, now time.Time, lagReco
 	// - MaxIncrease is too low and the rate might not increase
 	// - after the new rate was limited by the bad rate, we got the old rate
 	//   (In this case we might slightly go above the bad rate which we accept.)
-	if rate <= float64(oldRate) {
+	if int64(rate) <= oldRate {
 		rate = float64(oldRate) + memoryGranularity
 		increaseReason += fmt.Sprintf(" (minimum progress by %v)", memoryGranularity)
 		previousRateSource = "previous set rate"
